+++ api/src/main/java/org/allaymc/api/math/MathUtils.java
@@ -28,11 +28,12 @@ public final class MathUtils {
         throw new UnsupportedOperationException("This is a utility class and cannot be instantiated");
     }
 
+    // TODO: remove it (don't forget .max() method!)
+
     /**
      * Convert cloudburst vector to joml vector.
      *
      * @param cbVec cloudburst vector
-     *
      * @return joml vector, or {@code null} if the input is {@code null}
      */
     public static Vector3ic CBVecToJOMLVec(org.cloudburstmc.math.vector.Vector3i cbVec) {
@@ -47,7 +48,6 @@ public final class MathUtils {
      * Convert cloudburst vector to joml vector.
      *
      * @param cbVec cloudburst vector
-     *
      * @return joml vector, or {@code null} if the input is {@code null}
      */
     public static Vector3dc CBVecToJOMLVec(org.cloudburstmc.math.vector.Vector3d cbVec) {
@@ -62,7 +62,6 @@ public final class MathUtils {
      * Convert cloudburst vector to joml vector.
      *
      * @param cbVec cloudburst vector
-     *
      * @return joml vector, or {@code null} if the input is {@code null}
      */
     public static Vector3fc CBVecToJOMLVec(org.cloudburstmc.math.vector.Vector3f cbVec) {
@@ -77,7 +76,6 @@ public final class MathUtils {
      * Convert joml vector to cloudburst vector.
      *
      * @param JOMLVec joml vector
-     *
      * @return cloudburst vector, or {@code null} if the input is {@code null}
      */
     public static org.cloudburstmc.math.vector.Vector3i JOMLVecToCBVec(Vector3ic JOMLVec) {
@@ -92,7 +90,6 @@ public final class MathUtils {
      * Convert joml vector to cloudburst vector.
      *
      * @param JOMLVec joml vector
-     *
      * @return cloudburst vector, or {@code null} if the input is {@code null}
      */
     public static org.cloudburstmc.math.vector.Vector3d JOMLVecToCBVec(Vector3dc JOMLVec) {
@@ -107,7 +104,6 @@ public final class MathUtils {
      * Convert joml vector to cloudburst vector.
      *
      * @param JOMLVec joml vector
-     *
      * @return cloudburst vector, or {@code null} if the input is {@code null}
      */
     public static org.cloudburstmc.math.vector.Vector3f JOMLVecToCBVec(Vector3fc JOMLVec) {
@@ -122,7 +118,6 @@ public final class MathUtils {
      * Floor a vector.
      *
      * @param vector3d the vector to floor
-     *
      * @return the floored vector
      */
     public static Vector3i floor(Vector3dc vector3d) {
@@ -133,7 +128,6 @@ public final class MathUtils {
      * Floor a vector.
      *
      * @param vector3f the vector to floor
-     *
      * @return the floored vector
      */
     public static Vector3i floor(Vector3fc vector3f) {
@@ -144,7 +138,6 @@ public final class MathUtils {
      * Ceil a vector.
      *
      * @param vector3d the vector to ceil
-     *
      * @return the ceiled vector
      */
     public static Vector3i ceil(Vector3dc vector3d) {
@@ -155,7 +148,6 @@ public final class MathUtils {
      * Ceil a vector.
      *
      * @param vector3f the vector to ceil
-     *
      * @return the ceiled vector
      */
     public static Vector3i ceil(Vector3fc vector3f) {
@@ -166,7 +158,6 @@ public final class MathUtils {
      * Center a vector (Adding 0.5 to the coordinates of the vector).
      *
      * @param vector3i the vector to center
-     *
      * @return the centered vector
      */
     public static Vector3d center(Vector3ic vector3i) {
@@ -178,7 +169,6 @@ public final class MathUtils {
      *
      * @param d         the double to round
      * @param precision the precision
-     *
      * @return the rounded double
      */
     public static double round(double d, int precision) {
@@ -190,7 +180,6 @@ public final class MathUtils {
      * Fast inverse square root (float).
      *
      * @param x the value
-     *
      * @return the result
      */
     public static float fastFloatInverseSqrt(float x) {
@@ -206,7 +195,6 @@ public final class MathUtils {
      * Fast inverse square root (double).
      *
      * @param x the value
-     *
      * @return the result
      */
     public static double fastDoubleInverseSqrt(double x) {
@@ -224,7 +212,6 @@ public final class MathUtils {
      * @param l     the left
      * @param value the value to check
      * @param r     the right
-     *
      * @return {@code true} if the value is in the range, otherwise {@code false}.
      */
     public static boolean isInRange(double l, double value, double r) {
@@ -243,7 +230,6 @@ public final class MathUtils {
      *
      * @param yaw   the yaw
      * @param pitch the pitch
-     *
      * @return the direction vector
      */
     public static Vector3d getDirectionVector(double yaw, double pitch) {
@@ -259,7 +245,6 @@ public final class MathUtils {
      * Calculate yaw from the direction vector.
      *
      * @param vector the direction vector
-     *
      * @return the yaw
      */
     public static double getYawFromVector(Vector3dc vector) {
@@ -276,7 +261,6 @@ public final class MathUtils {
      * Calculate the pitch by the direction vector.
      *
      * @param vector the direction vector
-     *
      * @return the pitch
      */
     public static double getPitchFromVector(Vector3dc vector) {
@@ -296,7 +280,6 @@ public final class MathUtils {
      * Calculate sin value quickly by looking up a pre-calculated table.
      *
      * @param radian the radian
-     *
      * @return the sin value
      */
     public static double fastSin(float radian) {
@@ -307,7 +290,6 @@ public final class MathUtils {
      * Calculate sin value quickly by looking up a pre-calculated table.
      *
      * @param radian the radian
-     *
      * @return the sin value
      */
     public static double fastSin(double radian) {
@@ -318,7 +300,6 @@ public final class MathUtils {
      * Calculate cos value quickly by looking up a pre-calculated table.
      *
      * @param radian the radian
-     *
      * @return the cos value
      */
     public static double fastCos(float radian) {
@@ -329,7 +310,6 @@ public final class MathUtils {
      * Calculate cos value quickly by looking up a pre-calculated table.
      *
      * @param radian the radian
-     *
      * @return the cos value
      */
     public static double fastCos(double radian) {
@@ -340,7 +320,6 @@ public final class MathUtils {
      * Check if the vector contains NaN.
      *
      * @param v the vector to check
-     *
      * @return {@code true} if the vector contains NaN, {@code false} otherwise.
      */
     public static boolean hasNaN(Vector3fc v) {
@@ -351,7 +330,6 @@ public final class MathUtils {
      * Check if the vector contains NaN.
      *
      * @param v the vector to check
-     *
      * @return {@code true} if the vector contains NaN, {@code false} otherwise.
      */
     public static boolean hasNaN(Vector3dc v) {
@@ -362,7 +340,6 @@ public final class MathUtils {
      * Check if the location contains NaN.
      *
      * @param l the location to check
-     *
      * @return {@code true} if the vector contains NaN, {@code false} otherwise.
      */
     public static boolean hasNaN(Location3fc l) {
@@ -374,7 +351,6 @@ public final class MathUtils {
      * Check if the location contains NaN.
      *
      * @param l the location to check
-     *
      * @return {@code true} if the vector contains NaN, {@code false} otherwise.
      */
     public static boolean hasNaN(Location3dc l) {
@@ -388,7 +364,6 @@ public final class MathUtils {
      * If the vector is zero, it can't be normalized, otherwise a vector with three NaN values will be produced.
      *
      * @param v the vector
-     *
      * @return the normalized vector
      */
     public static Vector3f normalizeIfNotZero(Vector3f v) {
@@ -401,7 +376,6 @@ public final class MathUtils {
      * If the vector is zero, it can't be normalized, otherwise a vector with three NaN values will be produced.
      *
      * @param v the vector
-     *
      * @return the normalized vector
      */
     public static Vector3d normalizeIfNotZero(Vector3d v) {
@@ -415,7 +389,6 @@ public final class MathUtils {
      *
      * @param aabb   the AABB to grow
      * @param growth the vector
-     *
      * @return the grown AABB
      */
     public static AABBd grow(AABBd aabb, Vector3dc growth) {
@@ -429,30 +402,12 @@ public final class MathUtils {
     }
 
     /**
-     * Get the maximum value from an int array.
-     *
-     * @param array the int array to get the maximum value from
-     *
-     * @return the maximum int value
-     */
-    public static int max(int[] array) {
-        int max = Integer.MIN_VALUE;
-        for (int i : array) {
-            if (i > max) {
-                max = i;
-            }
-        }
-        return max;
-    }
-
-    /**
      * Calculate the Chebyshev distance between two points in a 2D space.
      *
      * @param x1 the x coordinate of the first point
      * @param y1 the y coordinate of the first point
      * @param x2 the x coordinate of the second point
      * @param y2 the y coordinate of the second point
-     *
      * @return the Chebyshev distance between the two points
      */
     public static int calculateChebyshevDistance(int x1, int y1, int x2, int y2) {
@@ -466,7 +421,6 @@ public final class MathUtils {
      * @param y1 the y coordinate of the first point
      * @param x2 the x coordinate of the second point
      * @param y2 the y coordinate of the second point
-     *
      * @return the Manhattan distance between the two points
      */
     public static int calculateManhattanDistance(int x1, int y1, int x2, int y2) {
@@ -480,7 +434,6 @@ public final class MathUtils {
      * @param y1 the y coordinate of the first point
      * @param x2 the x coordinate of the second point
      * @param y2 the y coordinate of the second point
-     *
      * @return the Euclidean distance between the two points
      */
     public static double calculateEuclideanDistance(int x1, int y1, int x2, int y2) {