+++ server/src/main/java/org/allaymc/server/command/tree/node/DifficultyNode.java
@@ -8,10 +8,10 @@ import org.allaymc.api.world.data.Difficulty;
  */
 public class DifficultyNode extends EnumNode {
     private static final String[] VALUES = new String[]{
-            "peaceful", "p", "0",
-            "easy", "e", "1",
-            "normal", "n", "2",
-            "hard", "h", "3"
+            "0", "peaceful", "p",
+            "1", "easy", "e",
+            "2", "normal", "n",
+            "3", "hard", "h"
     };
 
     public DifficultyNode(String name, CommandNode parent, Difficulty defaultValue) {
@@ -20,10 +20,12 @@ public class DifficultyNode extends EnumNode {
 
     @Override
     protected Difficulty argToResult(String arg) {
-        try {
-            return Difficulty.from(Integer.parseInt(arg));
-        } catch (NumberFormatException ignored) {
-            return Difficulty.from(arg);
-        }
+        return switch (arg) {
+            case "0", "peaceful", "p" -> Difficulty.PEACEFUL;
+            case "1", "easy", "e" -> Difficulty.EASY;
+            case "2", "normal", "n" -> Difficulty.NORMAL;
+            case "3", "hard", "h" -> Difficulty.HARD;
+            default -> null;
+        };
     }
 }